Added a replace in file method.
Now you can replace a single value in a file! :D This is the last change I'll do to file I/O for now, I'm gonna be working on the network system and in bettering the components management to try and stop using pointers. The networking system development might be slow though, I have a lot to learn about that.