TCP works
Now TCP is fully functional, I still need to close sockets when the server and client connection is lost but that's easy.
It wasn't working properly because the arguments for the network component constructor had the same name as the class variables, so it was storing garbage and so, the SDLNet couldn't work properly.

TODO: UDP