UDP Works, for real this time.
The problem had to do with me not using SDLNet_AllockPacket and FreePacket, that made the program crash randomly.

Other problem was when clearing the data from the framework I wouldn't check if the udp socket was null or not,  so it would try to close a null socket, making the program crash too.

It's all fixed now though, so I can now make online games :D

I still have to add some security methods for udp and tcp, but that won't be a problem, its just having a checking string in the server and when a client connects it has to send that string to be able to connect.