Fixed the last known segmentation fault
Now everything should be working as intended without memory issues. And now that I have those problems out of the way, I'll be working on screen effects again. Also, since now the ScriptsInterface holds the scripts, it should be, in theory, easier to create and load savestates. I'm going to try